+++ Northwind/Models/Request/Order/UpdateOrderRequest.cs
@@ -13,7 +13,7 @@ namespace Northwind.Models.Request.Order
         public DateTime? RequiredDate { get; set; }
         public DateTime? ShippedDate { get; set; }
         public int ShipViaId { get; set; }
-        public float Freight { get; set; }
+        public decimal Freight { get; set; }
         public string ShipName { get; set; }
         public string ShipAddress { get; set; }
         public string ShipCity { get; set; }
